IRoomConnectionInfo
