+++ Runtime/Scripts/Rooms/IRoomConnectionInfo.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3b99d13e69c14eeab9f90e335222d81d
+timeCreated: 1707480332
